fix the issue - disconnected after 1 hour

+++ Chart.cpp
@@ -12,7 +12,7 @@
 
 int MainWindow::encoderround1;    //catch the data from MainWindow.
 double MainWindow::motoround1;
-int MainWindow::biground1;
+double MainWindow::biground1;
 
 Chart::Chart(QWidget *parent) :
   QWidget(parent),
